+++ examples/nextjs/pages/index.tsx
@@ -1,61 +1,47 @@
 import type { NextPage } from 'next'
 import { useState } from 'react'
 
+import { Button, Container, Input, Title } from '@mantine/core'
 import {
   useAccessToken,
   useAuthenticated,
   useChangeEmail,
   useChangePassword,
-  useSignInEmailPassword,
-  useSignInEmailPasswordless,
-  useSignOut,
-  useSignUpEmailPassword
+  useSignOut
 } from '@nhost/nextjs'
 import { useAuthQuery } from '@nhost/react-apollo'
 
+import { authProtected } from '../components/protected-route'
 import { BOOKS_QUERY } from '../helpers'
 
 // * Reference: https://blog.codepen.io/2021/09/01/331-next-js-apollo-server-side-rendering-ssr/
 
 const Home: NextPage = () => {
   const isAuthenticated = useAuthenticated()
-  const [email, setEmail] = useState('')
-  const [password, setPassword] = useState('')
+  const [email] = useState('')
+  const [password] = useState('')
   const [newEmail, setNewEmail] = useState('')
   const [newPassword, setNewPassword] = useState('')
   const accessToken = useAccessToken()
   const { signOut } = useSignOut()
-  const { signUpEmailPassword, ...signUpResult } = useSignUpEmailPassword()
-  const { signInEmailPassword } = useSignInEmailPassword()
-  const { signInEmailPasswordless } = useSignInEmailPasswordless({ redirectTo: '/third' })
   const { changeEmail, ...changeEmailResult } = useChangeEmail()
   const { changePassword, ...changePasswordResult } = useChangePassword()
   const { loading, data, error } = useAuthQuery(BOOKS_QUERY)
   return (
-    <div>
+    <Container>
+      <Title>Index page</Title>
       {isAuthenticated ? (
         <>
-          <button onClick={signOut}>Logout</button>
-          <input value={newEmail} onChange={(e) => setNewEmail(e.target.value)} />
-          <button onClick={() => changeEmail(email)}>Change email</button>
+          <Button onClick={signOut}>Logout</Button>
+          <Input value={newEmail} onChange={(e) => setNewEmail(e.target.value)} />
+          <Button onClick={() => changeEmail(email)}>Change email</Button>
           <div>{JSON.stringify(changeEmailResult)}</div>
-          <button onClick={() => changePassword(password)}>Change password</button>
-          <input value={newPassword} onChange={(e) => setNewPassword(e.target.value)} />
+          <Button onClick={() => changePassword(password)}>Change password</Button>
+          <Input value={newPassword} onChange={(e) => setNewPassword(e.target.value)} />
           <div>{JSON.stringify(changePasswordResult)}</div>
         </>
       ) : (
-        <>
-          <input value={email} onChange={(e) => setEmail(e.target.value)} />
-          <button onClick={() => signInEmailPasswordless(email)}>Passwordless signin</button>
-          <div>{JSON.stringify(signUpResult)}</div>
-          <input value={password} onChange={(e) => setPassword(e.target.value)} type="password" />
-          <button onClick={() => signUpEmailPassword(email, password)}>
-            Email + password sign-up
-          </button>
-          <button onClick={() => signInEmailPassword(email, password)}>
-            Email + password sign-in
-          </button>
-        </>
+        <div>go to /sign-in</div>
       )}
 
       <p>Access Token</p>
@@ -68,8 +54,8 @@ const Home: NextPage = () => {
         </ul>
       )}
       {!loading && error && <div>ok {JSON.stringify(error)}</div>}
-    </div>
+    </Container>
   )
 }
 
-export default Home
+export default authProtected(Home)