+++ examples/nextjs/pages/sign-in/index.tsx
@@ -1,0 +1,26 @@
+import { NextPage } from 'next'
+import React from 'react'
+import { FaLock } from 'react-icons/fa'
+
+import { Divider } from '@mantine/core'
+
+import AuthLink from '../../components/AuthLink'
+import OAuthLinks from '../../components/OauthLinks'
+import SignInLayout from '../../layouts/SignInLayout'
+
+export const SignInPage: NextPage = () => {
+  return (
+    <SignInLayout title="Sign In">
+      <OAuthLinks />
+      <Divider my="sm" />
+      <AuthLink icon={<FaLock />} variant="outline" link="/sign-in/passwordless">
+        Continue with passwordless email
+      </AuthLink>
+      <AuthLink variant="subtle" link="/sign-in/email-password">
+        Continue with email + password
+      </AuthLink>
+    </SignInLayout>
+  )
+}
+
+export default SignInPage